temporal-io/my-python/MY/2_intermediate/2_hr_approval_slack/pyproject.toml
[project]
name = "hr-approval-slack"
version = "0.1.0"
description = "HR approval workflow with Slack integration demonstrating Temporal Updates"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "temporalio>=1.6.0",
    "slack-sdk>=3.33.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = []
